Add new bash scripts and Makefile tag rule
Added a new Makefile rule "tag" which executes "create-new-tag.sh" bash script. The bash script selects a new tag, and the Makefile rule sets this as tag using Git annotate. If no tag is provided, or the "select-new-tag.sh" is not executable, it will exit with an error. This update streamlines the process of creating and annotating new version tags.